Split router handlers into separate files and remove `/api` prefix
