+++ http-practice/routes/links/index.js
@@ -1,0 +1,11 @@
+const create = require('./create');
+const read = require('./read');
+const update = require('./update');
+const del = require('./delete');
+
+module.exports = {
+  create,
+  read,
+  update,
+  delete: del
+}
