+++ http-practice/scratch/index.html
@@ -9,7 +9,7 @@
   </head>
   <body>
     <div class="container">
-      <form action="http://localhost:3000/api/links/" method="POST">
+      <form action="http://localhost:3000/links/" method="POST">
         <h1>Shorten your URL</h1>
         <div class="form-control">
           <label for="">Link</label>
